Add files via upload
The objective of this use case is to perform sentiment analysis on the reviews of the All-New Fire HD 8 Tablet to understand customer sentiments towards the product. By analyzing these reviews, we aim to identify the overall customer satisfaction and key factors influencing positive and negative sentiments. This analysis will help Amazon and potential buyers gain insights into the strengths and weaknesses of the product. The results can also guide improvements in future product development and marketing strategies.